PEJT-29: change entity and other

+++ src/main/kotlin/io/aimc/reportingservice/dto/ReportDto.kt
@@ -4,6 +4,6 @@ import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlRootElement
 
 @JacksonXmlRootElement
 data class ReportDto(
-    val portionAmount: Int,
-    val shipmentAmount: Int
+    val portionAmount: Long,
+    val shipmentAmount: Long
 )
