+++ src/main/kotlin/io/aimc/reportingservice/model/Report.kt
@@ -1,6 +1,6 @@
 package io.aimc.reportingservice.model
 
 data class Report(
-    val portionAmount: Int,
-    val shipmentAmount: Int
+    val portionAmount: Long,
+    val shipmentAmount: Long
 )
